reuse StringStreamDataWriter
